feat: coinbase and pay-to-pubkey-hash

+++ src/config/models.rs
@@ -5,7 +5,7 @@ pub struct MiningConfig {
     pub block_time_secs: usize,
     pub block_adjustment_interval: usize,
     pub start_difficulty_bit: u8,
-    pub mining_reward: usize,
+    pub mining_reward: u64,
     pub reward_halvening_interval: usize,
 }
